# TestFragansa/TestFrag/bin/Release/net6.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "WebApi",
    "version": "v1"
  },
  "paths": {
    "/api/Actores": {
      "get": {
        "tags": [
          "Actor"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Actores"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Actor"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Actores"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Actor"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Actores"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Actores/{Id}": {
      "get": {
        "tags": [
          "Actor"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "Id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Actores"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Actor"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "Id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Autenticacion/Validar": {
      "post": {
        "tags": [
          "Autenticacion"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Usuarios"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Director": {
      "get": {
        "tags": [
          "Director"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Directores"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Director"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Directores"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Director"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Directores"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Director/{Id}": {
      "get": {
        "tags": [
          "Director"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "Id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Directores"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Director"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "Id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Genero": {
      "get": {
        "tags": [
          "Genero"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Generos"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Genero"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Generos"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Genero"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Generos"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Genero/{Id}": {
      "get": {
        "tags": [
          "Genero"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "Id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Generos"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Genero"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "Id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Pais": {
      "get": {
        "tags": [
          "Pais"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Paises"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Pais"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Paises"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Pais"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Paises"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Pais/{Id}": {
      "get": {
        "tags": [
          "Pais"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "Id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Paises"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Pais"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "Id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Peliculas": {
      "get": {
        "tags": [
          "Pelicula"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Peliculas"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Pelicula"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Peliculas"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Pelicula"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Peliculas"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Peliculas/{Id}": {
      "get": {
        "tags": [
          "Pelicula"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "Id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Peliculas"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Pelicula"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "Id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "definitions": {
    "Actores": {
      "type": "object",
      "properties": {
        "idActor": {
          "format": "int32",
          "type": "integer"
        },
        "nombre": {
          "type": "string"
        },
        "apellido": {
          "type": "string"
        },
        "pais": {
          "format": "int32",
          "type": "integer"
        },
        "paises": {
          "$ref": "#/definitions/Paises"
        }
      }
    },
    "Directores": {
      "type": "object",
      "properties": {
        "idDirector": {
          "format": "int32",
          "type": "integer"
        },
        "nombre": {
          "type": "string"
        },
        "apellido": {
          "type": "string"
        },
        "idPais": {
          "format": "int32",
          "type": "integer"
        },
        "paises": {
          "$ref": "#/definitions/Paises"
        }
      }
    },
    "Generos": {
      "type": "object",
      "properties": {
        "idGenero": {
          "format": "int32",
          "type": "integer"
        },
        "nombre": {
          "type": "string"
        }
      }
    },
    "Paises": {
      "type": "object",
      "properties": {
        "idPais": {
          "format": "int32",
          "type": "integer"
        },
        "nombre": {
          "type": "string"
        }
      }
    },
    "Peliculas": {
      "type": "object",
      "properties": {
        "idPelicula": {
          "format": "int32",
          "type": "integer"
        },
        "genero": {
          "format": "int32",
          "type": "integer"
        },
        "pais": {
          "format": "int32",
          "type": "integer"
        },
        "actor": {
          "format": "int32",
          "type": "integer"
        },
        "director": {
          "format": "int32",
          "type": "integer"
        },
        "titulo": {
          "type": "string"
        },
        "reseña": {
          "type": "string"
        },
        "imagenUrl": {
          "type": "string"
        },
        "codigoTrailer": {
          "type": "string"
        }
      }
    },
    "Usuarios": {
      "type": "object",
      "properties": {
        "idUsuario": {
          "format": "int32",
          "type": "integer"
        },
        "usuario": {
          "type": "string"
        },
        "clave": {
          "type": "string"
        }
      }
    }
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}